Fix problem introduced by 2e556603 on 64bit systems
Use "unsigned ints" instead of size_t to avoid 64 bit boxed integers.

+++ bytecode.lua
@@ -193,8 +193,8 @@ ffi.cdef[[
     int free(void*);
 
     typedef struct Buf {
-        size_t size;
-        size_t offs;
+        unsigned int size;
+        unsigned int offs;
         uint8_t *data;
     } Buf;
 ]]
